When a data loader fails, unlink any previously cached output.
To test, open a page that has a working data loader, then edit the loader into an error state; the page should display an "Error: Unable to load file", instead of the older data.

Implementation note: we're not sure there is an old file at the outputPath, but we don't check, and don't await, which ignores any error.