save point, WIP for #423
`bundleStyles` correctly puts the `@import "url";` lines at the top of its text. In the preview or build scripts, we can easily detect these lines with RegExp and maybe remove them from the bundled CSS—though that is not even strictly necessary.

Hoisting them as stand-alone links is what will give the performance boost, and it's a bit more difficult, because in both cases (preview and render) we bundle much later than we link, so it looks like we have to change the order of operations. I think it's doable but I'm not fully there yet.