Increase the required version of send from 0.0.18 to 0.0.19
The version of send 0.0.18 has a published vulnerability that is fixed in
0.0.19, see https://github.com/advisories/GHSA-m6fv-jmcg-4jfg. However, the
version specifier of "^0.0.18" in package.json does not allow projects to use
0.0.19.

This changes the specifier to "^0.0.19" which allows (and requires) Observable
Framework projects to use 0.0.19 to avoid the vulnerability.

Note that the vulnerability https://github.com/advisories/GHSA-m6fv-jmcg-4jfg
has been public for over 2 weeks at this point, and so this isn't disclosing
anything new.