add lastModified to a module’s hash when it references a file attachment (#1837)
* when a module references a file attachment, include the file's lastModified timestamp in the module's hash.

closes #1836

Note that this allows us to remove from the tests the ts rewrite that we used to do on modules, because their payload’s lastModified is now always set to the fake currentDate injected by the tests. Some tests don't want to use the fake date though! (I hope they never run in parallel?)

* clear currentDate in after

* prettier

---------

Co-authored-by: Mike Bostock <mbostock@gmail.com>